requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg==3.2.3
pydantic-settings==2.5.2
alembic==1.13.2
python-dotenv==1.0.1
pg8000==1.29.8
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
httpx==0.27.0
email-validator==2.1.0